Double click on an entity selects all entities of that type that are on screen #22
Instead decided to add all units with the same class within a range as it's easier to implement and does basically the same thing but with less processing.

The original idea meant we had to calculate the screen space for the box we were selecting in select the actors within that box. By just selecting within a range of the actor we just need to iterate through the actors and check their range, it also has the benefit if that we don't select actors that are off in the distance if the camera is at a low angle